Moved some functions to channel_service
